Fix pyenv path issue when building Lambda packages via Terraform
Ran into an interesting issue as I was upgrading our rsspca version in the platform-services-core project. Since it's now using a more recent version of the ubuntu machine, the following error occured:

https://app.circleci.com/pipelines/github/RSS-Engineering/platform-services-core/2096/workflows/e71d14e1-149c-4b72-b5b7-2b7377de48f5/jobs/58872

Found this as a solution after googling:

https://github.com/CircleCI-Public/gcp-cli-orb/issues/22

After changing the deploy job to use single quotes when exporting the PATH, the job was green again.