Artifact-based deployments
I have noticed that some of our more complex CircleCI workflows had been suboptimal for various reasons.

One of them was that in some cases we ended up building Lambda packages twice: once with the SAM build step, then during deploy if the Terraform Lambda module was used with certain parameters.

The other issue was attaching the workspace to deploy jobs that were behind a manual hold step, because at that point every single job prior to the approval step became a parent of the subsequent deploy jobs, which meant that every single workspace was attached by CircleCI and that operation can be quite slow.